fix(nutrition): targetCalories Math.round + burnedCalories int 타입 일관성
- (int) goal.getTargetValue() → Math.round() 적용으로 반올림 처리
- CalorieBalanceResponse의 burnedCalories를 long → int로 변경
- 서비스에서 캐스팅 추가: (int) routineSessionRepository...
- 테스트 어설션 int 타입으로 통일

+++ backend/src/main/java/com/ssafy/manager/nutrition/presentation/dto/CalorieBalanceResponse.java
@@ -3,7 +3,7 @@ package com.ssafy.manager.nutrition.presentation.dto;
 public record CalorieBalanceResponse(
         int targetCalories,
         double intakeCalories,
-        long burnedCalories,
+        int burnedCalories,
         double remainingCalories,
         int mealCount,
         boolean lastMealRecommendTrigger
